refactor: 파일 구조 변경 및 코드 리팩토링
