+++ backend/src/main/java/com/woori/wonfit/member/liked/domain/Liked.java
@@ -1,7 +1,6 @@
 package com.woori.wonfit.member.liked.domain;
 
 import com.woori.wonfit.member.member.domain.Member;
-import com.woori.wonfit.product.product.domain.Product;
 import lombok.Getter;
 
 import javax.persistence.*;
@@ -13,10 +12,6 @@ public class Liked {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
-//    @ManyToOne
-//    @JoinColumn(name = "product_id")
-//    private Product product;
 
     @ManyToOne
     @JoinColumn(name = "member_id")
