+++ backend/src/main/java/com/woori/wonfit/member/member/repository/MemberRepository.java
@@ -1,4 +1,4 @@
-package com.woori.wonfit.member.member;
+package com.woori.wonfit.member.member.repository;
 
 
 import com.woori.wonfit.member.member.domain.Member;
